// src/app/shared/model/ref-employee-obj.ts
export class RefEmployeeObj{
    RefEmpId: number;
    EmpNo: string;
    EmpName: string;
    JoinDt: any;
    Addr: string;
    Zipcode: string;
    AreaCode1: string;
    AreaCode2: string;
    AreaCode3: string;
    AreaCode4: string;
    City: string;
    PhnArea1: string;
    Phn1: string;
    PhnExt1: string;
    PhnArea2: string;
    Phn2: string;
    PhnExt2: string;
    PhnArea3: string;
    Phn3: string;
    PhnExt3: string;
    FaxArea: string;
    Fax: string;
    MobilePhnNo1: string;
    MobilePhnNo2: string;
    Email1: string;
    Email2: string;
    IsExt: boolean;
    TaxIdNo: string;
    MrIdTypeCode: string;
    IdNo: string;
    ImageLocation: string;
    Loginsoftphone: string;
    IsLeave: boolean;
    IsActive: boolean;
    RowVersion: string;
    
    constructor() { this.RefEmpId = 0, this.RowVersion = "" }
}