// src/app/shared/model/office-obj.model.ts
// Generated by https://quicktype.io

export class OfficeObj {
    RefOfficeId: number;
    OfficeCode: string;
    OfficeShortName: string;
    OfficeName: string;
    OrgMdlId: number;
    OfficeAddr: string;
    AreaCode1: string;
    AreaCode2: string;
    AreaCode3: string;
    AreaCode4: string;
    City: string;
    ZipCode: string;
    PhnArea1: string;
    Phn1: string;
    PhnExt1:string;
    PhnArea2: string;
    Phn2: string;
    PhnExt2:string;
    FaxArea: string;
    Fax: string;
    CntctPersonName: string;
    CntctPersonJobTitle: string;
    CntctPersonMobilePhnNo1: string;
    CntctPersonMobilePhnNo2: string;
    MrOfficeClassCode: string;
    RefOfficeAreaId: number;
    IsActive: boolean;
    ParentId: number;
    IsOfficeClose: boolean;
    OfficeOpeningDt: Date;
    IsAllowAppCreated: boolean;
    HolidaySchmHId: number;
    CntctPersonEmail1: string;
    PhnArea3: string;
    Phn3: string;
    PhnExt3: string;
    CntctPersonEmail2: string;
    WorkingHourSchmHId: number;
    RefTaxOfficeId: number;
    IsVirtualOffice: boolean;
    MrKonvenSyariahCode: string;
    MrOfficeTypeCode:string;
    CenterGrpCode:string;
    CenterGrpName:string;
    OfficeParent: number;
    RowVersion: string;
    ParentOfficeCode: string;
    MrCenterGrpTypeCode: string;
    IsHaveCashier : boolean;
    HierarchyLvl : number;
    constructor() { this.RefOfficeId = 0, this.RowVersion = "" }
}